--- packages/collections/src/Adaptor.js
@@ -1,1 +1,0 @@
-export default {};
